Backend/requirements.txt
# FastAPI and Server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.17

# Database
sqlalchemy>=2.0.36
pg8000>=1.31.2
alembic>=1.14.0

# Security
passlib>=1.7.4
python-jose[cryptography]>=3.3.0
bcrypt>=4.2.1
pyotp>=2.9.0

# Validation
pydantic>=2.10.0
pydantic-settings>=2.6.0
email-validator>=2.2.0

# OAuth
authlib>=1.4.0
httpx>=0.28.0
itsdangerous>=2.2.0

# Email - use standard smtplib (built-in)
# aiosmtplib removed due to Python 3.13 compatibility

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0

# Utilities
python-dotenv>=1.0.1
